+++ CHANGELOG.md
@@ -12,6 +12,8 @@
   The 100 limit probably needs to be configurable.
 * Added register profiles to show only interesting/relevant registers.
 * Added UTF-8,16,32 support in the Memory Visualizer.
+* Added an internal "dark" and "light" theme via the View->Style menu.
+* Added option to open the current source file in an external editor.
 
 ## [2.3] - 2023-11-19
 * In the margins of the source windows, allow CTRL+DoubleClick to do a quick RunToLine or RunToAddress.